Create Todo_list.html
This is a simple Todo List App implemented using HTML, CSS, and JavaScript. The app allows users to add tasks to their todo list, mark them as complete by checking the checkboxes, and apply a strikethrough effect to the completed tasks. The app also displays the current date at the top.

Users can enter new tasks using the input field and click the "+" button to add them to the list dynamically. The app supports adding an unlimited number of tasks.

The "Clear All" button allows users to remove all the tasks from the list at once.

The app is built using the Bootstrap framework for styling and utilizes event listeners and DOM manipulation to handle user interactions.

Please note that this code snippet is intended to be used within an HTML file and includes references to external resources such as the Bootstrap CSS and JavaScript libraries.